Adds IsZero functionality to DateTime
This PR adds a new function `IsZero` to the DateTime struct. The goal of
this is to satisfy a common interface, IsZeroer, used by go-yaml/yaml
marshaller. `IsZero` already exists on the time.Time object, so this
does not introduce any novel implementation.

For reference, here is the [documentation
note](https://pkg.go.dev/gopkg.in/yaml.v3#Marshal) on using IsZero by the
`(yaml).Marshal` function.

Signed-off-by: Brandon Gonzalez <bg@chronosphere.io>