chore(deps): removed dependency to govalidator
github.com/asaskevich/govalidator hasn't evolved in the recent years.

It has proved a useful dependency so far, but now we prefer to duplicate
the few lines of codes that we were still using so we are more in
control of the validations being carried out.

In particular, we may reduce further the footprint of regexp-based
validations.

Signed-off-by: Frederic BIDON <fredbi@yahoo.com>